design of a simple and common interface to all analysis.
this interface is simple because there is not so much in
common when we consider the different analysis we have implemented.
in particular, one of the analysis is intraprocedural (and we instantiate
one object of this analysis to each method). the other analysis
are "whole program analysis", and we instantiate only once for the
whole program.